refactor: 파일 이름 변경으로 인한 경로 수정
